refactor: move transposition logic to helper function
