+++ src/main.rs
@@ -2,7 +2,7 @@ use actix_web::{middleware::Logger, web, App, HttpServer};
 use clap::{Subcommand, Parser};
 use hound::{Error, WavReader, WavSpec, WavWriter};
 
-use audi8::{api::{health, transpose_wav}, error::InsufficientInputError, time_scaler::TimeScaler, wav};
+use audi8::{api::{health, transpose_wav}, error::InsufficientInputError, time_scaler::TimeScaler, transpose, wav};
 use rubato::{FftFixedInOut, Resampler};
 
 #[derive(Parser)]
@@ -50,8 +50,6 @@ async fn main() -> std::io::Result<()> {
                 Err(err) => panic!("could not open audio file: {err}")
             };
             let spec = reader.spec();
-            println!("spec: {spec:?}");
-
             let mut writer = WavWriter::create(
                 &output_file,
                 WavSpec {
@@ -62,46 +60,7 @@ async fn main() -> std::io::Result<()> {
                 },
             ).unwrap();
 
-            let chunk_size = 4096usize;
-            let sampling_ratio = 2f32.powf(num_semitones as f32 / 12.0);
-
-            let mut resampler = FftFixedInOut::<f32>::new(
-                spec.sample_rate as usize,
-                (spec.sample_rate as f32 / sampling_ratio) as usize,
-                chunk_size,
-                spec.channels as usize,
-            ).unwrap();
-
-            let mut time_scaler = TimeScaler::new(
-                chunk_size,
-                chunk_size / 2,
-                sampling_ratio,
-                spec.channels as usize,
-            );
-
-            // Read entire file into input buffer
-            loop {
-                let in_block = wav::read_frames(&mut reader, chunk_size);
-                let exhausted = in_block.iter().any(|channel| channel.len() < chunk_size);
-                time_scaler.push_block(&in_block);
-
-                if exhausted {
-                    break;
-                }
-            }
-
-            loop {
-                let in_frames = resampler.input_frames_next();
-                let out_frames = resampler.output_frames_next();
-                match time_scaler.pop_frames(in_frames) {
-                    Ok(block) => {
-                        let out_block = resampler.process(&block, None).unwrap();
-                        wav::write_frames(&mut writer, out_block, out_frames);
-                    }
-                    Err(InsufficientInputError) => break,
-                }
-
-            }
+            transpose::transpose_audio(&mut reader, &mut writer, num_semitones);
 
             writer.finalize().unwrap();
